third_party/WebKit/LayoutTests/fast/css/font-face-unicode-range-ligatures.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<style type="text/css">
    @font-face {
    font-family: libertinerangesplit;
    src: url("../../third_party/Libertine/LinLibertine_R.woff");
    unicode-range: U+0066, U+0069;
    }

    @font-face {
    font-family: libertinefull;
    src: url("../../third_party/Libertine/LinLibertine_R.woff");
    }

    .test {
    font-size: 200px;
    }

    #libertinerangesplit {
    font-family: libertinerangesplit;
    visibility: hidden;
    }

    #libertinefull {
    font-family: libertinefull;
    visibility: hidden;
    }
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script type="text/javascript">

async_test(function(testHandle) {
    document.fonts.ready.then( function() {
        testHandle.step( function() {
            rangesplitwidth = document.querySelector("#libertinerangesplit").getBoundingClientRect().width;
            fullwidth = document.querySelector("#libertinefull").getBoundingClientRect().width;
            assert_equals(rangesplitwidth, fullwidth);
        });
        testHandle.done();
    });
}, "Ligature for fi formed for full font and subsetted font.");

</script>
</head>
<body>
    <div class="test">
        <span id="libertinerangesplit">fi</span>
        <span id="libertinefull">fi</span>
    </div>
</body>
